Moving user-messages into “generic” sub-directory.
