+++ src/alerts/enrollment-alert/hooks.js
@@ -2,10 +2,9 @@
 import {
   useContext, useState, useCallback,
 } from 'react';
-import { UserMessagesContext, ALERT_TYPES, useAlert } from '../../user-messages';
+import { UserMessagesContext, ALERT_TYPES, useAlert } from '../../generic/user-messages';
 import { useModel } from '../../generic/model-store';
 import { postCourseEnrollment } from './data/api';
-
 
 export function useEnrollmentAlert(courseId) {
   const course = useModel('courses', courseId);
