+++ src/index.jsx
@@ -14,7 +14,7 @@ import { messages as headerMessages } from '@edx/frontend-component-header';
 import Footer, { messages as footerMessages } from '@edx/frontend-component-footer';
 
 import appMessages from './i18n';
-import { UserMessagesProvider } from './user-messages';
+import { UserMessagesProvider } from './generic/user-messages';
 
 import './index.scss';
 import './assets/favicon.ico';
